feat: Update dependencies and add Firebase, Hilt, Room, Retrofit, OkHttp, and Coil
This commit updates the dependencies in `gradle/libs.versions.toml` to newer versions, including:
- Compose BOM
- Firebase libraries
- Material and Material3
- Navigation Compose
- Retrofit and OkHttp
- Coroutines
- Coil
- Hilt
- Room
- UI

It also adds new dependencies to `gradle/libs.versions.toml`, including:
- Firebase BOM, Analytics, Auth and Firestore
- Material icons
- Room libraries
- Retrofit related libraries
- Okhttp logging interceptor
- Coroutines
- Coil related libraries
- Hilt related libraries
- Ui library
- Kotlin serialization

The `build.gradle.kts` and `app/build.gradle.kts` files are modified to apply the `firebase-services`, `dagger-hilt`, `kapt`, and `kotlin-serialization` plugins, and add the new dependencies to the project.

The `MainActivity.kt` is updated to import firebase related classes, and `AndroidManifest.xml` is modified to include the internet permission.